Add type safety when computing consul key paths
Previously the kp package provided functions (e.g.
kp.IntentPath(string...)) that would simply join the passed strings
with a "/" in between and return a consul path that would be used. This
means all callers need to know the structure of consul keys and do
sanity checking against bad things happening.

These functions are replaced with functions that are named more
explicitly to indicate what the caller wants.