gzip: Create parents
In our tar extraction code, if we were to encounter a tarball with a
file `a/b` but not the directory `a/`, we would attenpt to create `a/b`
without `a/` existing and fail in doing so.

This commit always creates the parent of the item we are about to
extract so that we avoid such problems.