gzip: Error on paths above destination
If a tarball contains a path of the form `../a`, we wouldn't want to
allow this to escape from the intended destination directory. Such
behavior is unsafe, as it might allow modification of system files.

Behavior of BSD tar: Forbids extracting such a tarball unless
specifically directed to do so with the `-P` flag.

Behavior of GNU tar: Strips the leading `../` from the path, such that
you get `./a` extracted instead.

We propose that silently changing the extraction target might surprise
the user more than simply raising an error, so we prefer to raise the
error.