gzip: Error on escaping symlinks
Consider a tarball with `a/b` being extracted to `/my/jail`. Now let's
consider the case where `/my/jail/a` is a symlink pointing to `..` or
even `/usr` (this may happen through a previous entry in the same
tarball, a previous tarball, or any variety of reasons).

If we do not check symlinks, we may blindly extract `b` to a path
outside of `/my/jail`. We should disallow this, as this sort of path
traversal can result in unsafe behavior (modification of system files).

Behavior of BSD tar: Forbids extracting such a tarball unless
specifically directed to do so with the `-P` flag.

Behavior of GNU tar: Extracts normally, such that you get `/my/b`
or `/usr/b` extracted, escaping the jail.

We will err on the side of caution because we may extract files as a
privileged user. Therefore we will raise an error in this case.