Launchables may specify environment variables
This is a useful feature that is mostly, but not entirely,
solved by CONFIG_PATH. There are occasions where code
cannot conveniently use the exported YAML file for config,
despite being a more flexible configuration system. This
is especially true for legacy or 3rd party software that
is conventionally configured with environment variables.

A bug I discovered while implementing this: the current
implementation of the opencontainer launchable does not
support pod-supplied variables.