Option to query health through localhost
Since p2-preparer runs on the same node, it can talk to the app on
localhost in order to check its health. This way apps can serve the
status port on the loopback interface only. In many cases, it's not
necessary to expose it to the outside world. A new status option,
localhost_only, can be used to tell p2-preparer to check status through
the loopback interface.