Add p2-rm command
This commit adds the "p2-rm" command, which can be used to remove/unschedule a
pod from a node. It will delete the pod's manifest from intent and remove any
labels attached to the pod. If the node pod is being managed by a controller,
it will print a warning message but allow the deletion.

Sample execution:

    $ p2-rm --node mynode mypod
    warning: mynode/mypod is managed by replication controller: 12345678
             To reduce the controller's capacity, use "p2-rctl set-replicas"
    mynode: successfully removed mypod
    $ p2-rm --node mynode unmanagedpod
    mynode: successfully removed unmanagedpod