Protect against panic due to non-existent RC
rcstore.Get() returned a fields.RC, and in the event a key is missing,
an empty struct is returned. This made it very easy for callers to panic
when doing operations on empty fields (e.g. rc.Manifest.ID()). This can
occur in a farm if an inconsistent rolling update refers to RCs that do
not exist.

Now rcstore.Get() returns an error of a special type when the key doesn't
exist, more clearly representing a missing key. All callers were changed
to check for this error value before further processing.