Bug fix: release pod locks on replication init errors
When a replication is initialized, it attempts to acquire pod locks on all the
pods it will be deploying to. A new kp.Lock object is created, and locks are
acquired sequentially. If any pod is already locked, the replication will
abort. The bug is that the Lock is leaked without being cleaned up, so any
locks acquired in the attempt will stay locked until a transient network error
happens or (more likely) the process is restarted.

This change explicitly destroys the lock when a replication error occurs
durining initialization. Once all locks have been acquired and a "replication"
object is returned, it's the caller's responsibility to ensure that it is
closed properly.