Rename kp.Lock to kp.Session.
The interface wraps consul sessions, which are acquired from the store
and then subsequently may be used to lock one or more keys. A "lock"
would more accurately describe a lock on a key obtained by a session.

Rename pkg/kp/lock.go to pkg/kp/session.go

Make Lock() on kp.Session return a kp.Unlocker

Previously, the caller acquiring a lock had to know the string key that
was used in the Lock() in order to unlock it. However, it may not always
make sense to expose this internal lock implementation to the caller,
e.g. when locking a resource like a replication controller.

Now, Lock() returns a kp.Unlocker interface that has an Unlock()
function, as well as a Key() function for callers that want to log the
key being locked.

Callers don't have to call Unlock() to release their locks, they can
also destroy the underlying session (which will also occur if the
session is not renewed within the LockTTL period).