Modify rollstore.Store interface to have 3 creation functions
There are a multitude of scenarios that an RU might be created in, each
with different methods of maintaining data integrity. Rather than try to
cram all of that functionality into one function, 3 separate functions
were added to the interface to be called in different situations. It
should be obvious to the caller which situation it is in.

The 3 situations are:
1) Creating an RU using existing RCs for old and new.
2) Creating an RU using an existing old RC with a known ID, but creating
the new RC anew.
3) Creating an RU and usingt  a label selector for the old RC. If no RCs
match the old RC selector, a "dummy" one matching the specifications of
the new RC will be created.