Only attempt to unlock an RC if the lock is held
This is a bug introduced by https://github.com/square/p2/pull/402, in
which the library that handles consul locking was changed. Previously
you could call session.Unlock(key) on a key that you may or may not
hold, and it would not panic the process.

Now, however, session.Lock(key) returns a kp.Unlocker object that is
used to unlock a key. The code fixed in this PR had assumed that it was
okay to call unlock on a key that wasn't held, but this resulted in a
panic if the unlocker was nil