Make *update implement Update (adding pointer receiver)
Some functions on the implementation of the Update interface need to
mutate the underlying struct, particularly to keep a kp.Unlocker handle
around to unlock a lock at a later stage of an update.

Previously these functions were defined on an update value, rather than
pointer, so any changes to the struct were not visible from other
functions. This would result in locks never being released, because the
u.oldRCUnlocker and u.newRCUnlocker fields would always be nil.