roll: Support capacity increases
Sometimes (particularly in a migration from another deployment system),
we perform a capacity increase in conjunction with a rolling update. We
can support such cases by detecting when the number of desired replicas
in total between the old and new RCs is less than the final number of
desired replicas. In such cases, it's safe to deploy the difference
betwen the final desire and the current desire sum.

Example 1:

* Old RC and New RC both have desired nodes == 0.
* Final number of nodes desired is 3.
* Minimum health is 2.

In this case, one node should be scheduled onto the new RC, leaving the
old RC untouched.

Example 2:

* Old RC has desired nodes == 3.
* New RC has desired nodes == 0.
* Final number of nodes desired is 4.
* Minimum health is 3.

In this case, at some point during the rollout a single node should be
scheduled onto the new RC, leaving the old RC untouched.

NOTE that if both of the following are true:

* A mutable deployment model (Old RC and New RC share nodes)
* Old manifest and new manifest are different

This may perform an unsafe operation if the new RC chooses to schedule
onto node occupied by the old RC before it has scheduled onto its new
node.

We should attempt to prevent such an operation, separately from this.