roll: Finish update when new healthy > minimum requested
This removes the old block for healthy nodes that was effectively a
reimplementation of slice divisors. This check is unnecessary because
`rollAlgorithm` already ensures safe operation. Removing it has the
benefit of allowing deploys to proceed with more parallelism - if one
node is slow to become healthy but there is health headroom, the deploy
can proceed to other nodes in the meantime. This was not possible with
the block in place.

The other benefit is that in the case of a deploy where minimum health
requested is zero, the deploy terminates as soon as all nodes have
converged, without waiting for nodes to become healthy. This is useful
in the case where we are knowingly deploying a bad sha, for which nodes
will never become healthy.