roll: conservatively treat old unknown-health nodes as healthy
We wish to be conservative because we don't want to mistakenly take down
a node of unknown health only to find out that it was healthy. The other
advantage of this is that it supports easily migrating apps from other
deployment systems which may not report health.

NOTE that we wouldn't want to treat NEW unknown-health nodes as healthy.
Again, this is the conservative option.

Imagine that because of a system-wide error, all node healths become
unknown. IF we were to consider all NEW unknown-health nodes as healthy,
we may falsely assume that the minimum health requirement has been
satisfied, and take down all old nodes, violating the minimum health
requirement specified. This, it is conservative to assume only OLD
unknown-health nodes are healthy.