Allow specifying new RC labels to apply before RU creation
Prior to commit ae763a8b748, it was the responsibility of callers into
the rollstore library to first create the RCs that a new RU will
operate on. ae763a8b748 aims to provide atomicity of RC creation and RU
creation so that it is unlikely for data to end up in an bad,
inconsistent state.

However, some callers depended on the ability to add labels to the new
RC before the RU object is created, so that custom handlers of RUs could
use the RC's labels to define behavior. This ability was lost.

Now, RC labels can be specified in the various RU creation functions on
rollstore so that this capability is preserved.