Replace rollstore.Put() function
This was the only operation provided in the past for creating rolling
updates. Its use is discouraged because it does not provide any safety
guarantees around consistency of data (e.g. a RU being created with
references to RCs which don't exist).

It is being resurrected to ease clients of this library onto the new
functions without forcing.