roll: If update is done but just waiting on healthy, don't schedule
We ran into a situation today where a deploy was in a state where one
canary node was desired, but two nodes had appeared on the new side.
Examination of the logs produced the following

    {
      "msg": "Adding 1 new nodes",
      "new": {
        "Desired": 0,
        "Current": 0,
        "Real": 0,
        "Healthy": 0
      },
      "next": 1,
      "old": {
        "Desired": 19,
        "Current": 19,
        "Real": 19,
        "Healthy": 18
      },
      "pod": "censored",
    },

This is expected.

    {
      "msg": "Adding 1 new nodes",
      "new": {
        "Desired": 1,
        "Current": 1,
        "Real": 1,
        "Healthy": 0
      },
      "next": 1,
      "old": {
        "Desired": 18,
        "Current": 19,
        "Real": 19,
        "Healthy": 18
      },
      "pod": "censored",
    }

This was unexpected, but (assuming minimum healthy was < 18)
understandable because in `rollAlgorithm`, `headroom` would be > 1 and
`remaining` would be 1, so we would schedule one node.

This is a direct consequence of 0768d697962d100bb4f2366d5e6e52a93ff24b8b
which removed the block for healthy nodes. It turns out that this block
is a good thing to have... when the deploy is almost complete ("almost
complete" meaning the desired number of nodes have been scheduled on the
new RC).

If the deploy is almost complete by this definition, we need to check
whether the minimum health requirement is met. If so, we declare the
deploy complete (old behavior unchanged by this commit). If not, we wait
for minimum health (behavior changed by this commit - behavior before
this commit was to schedule new nodes).