roll: terminate when OLD + new healthy exceed minimum health
This helps when we are in a canary deploy. Imagine a canary deploy with
2 minimum health but only 1 canary desired.

The RU will move 1 node to the new side, then wait for 2 nodes to be
healthy on the new side, which is impossible.

The old side is allowed to contribute toward health too. This will allow
the above-mentioned canary deploy to terminate.