Only write the latest RC list to consumer of rcstore.WatchNew
Previously, if the consumer of the RC watch was very slow, we might
receive new updates from consul while we're still blocking on writing
the previous update. Now, if we receive a new update while still
attempting to write the old one, we just move on to the most recent
update.