fix misleading log line that runs each RC loop
It used to say "meeting desires with <rc info>" but this is misleading
if you don't have context, because the desires may or may not actually
be met. The new wording is clearer that an update is being handled, we
don't actually know if any action needs to be taken or not.