Failsafes for rc farms
This PR adds two failsafes to the RC farms to prevent
catastrophic unscheduling of pods across the cluster. This
situation is generally hypothetical, but I'd rather be safe
than sorry.

The first situation is - the farm will panic if zero RCs have
been scheduled in Consul. This currently doesn't prevent any
specific behavior other than the farm abandoning working on RCs,
but is worth signaling that it has occurred.

The second situation would be genuinely catastrophic. If replica
counts were all 0, the RC farm would go on a killing spree across
the cluster, unscheduling all services. In practice this should
never happen.