roll: set remaining as target desire - new desire, not new healthy
We experienced the following with a deploy with desired 3 nodes and
minimum 1 node.

```json
{
  "msg": "Adding 2 new nodes and removing 2 old nodes",
  "new": {
    "Desired": 0,
    "Current": 0,
    "Real": 0,
    "Healthy": 0,
    "Unhealthy": 0,
    "Unknown": 0
  },
  "nextAdd": 2,
  "nextRemove": 2,
  "old": {
    "Desired": 3,
    "Current": 3,
    "Real": 3,
    "Healthy": 3,
    "Unhealthy": 0,
    "Unknown": 0
  },
  "pod": "censored",
},
{
  "msg": "Adding 2 new nodes and removing 2 old nodes",
  "new": {
    "Desired": 2,
    "Current": 2,
    "Real": 2,
    "Healthy": 1,
    "Unhealthy": 1,
    "Unknown": 0
  },
  "nextAdd": 2,
  "nextRemove": 2,
  "old": {
    "Desired": 1,
    "Current": 1,
    "Real": 1,
    "Healthy": 1,
    "Unhealthy": 0,
    "Unknown": 0
    },
  "pod": "censored",
}
```

The first is OK, but the second represents a serious error since now
there are four desired nodes which is greater than that requested. This
is because the minimum remaining count was being calculated as
`targetDesired - newHealthy`, which is incorrect if newHealthy hasn't
caught up with newDesired yet. Instead the correct calculation is
`targetDesired - newDesired`.

This was not visible before #414 and got merged, because a conditional
removed in that PR caused it to always be the case that newHealthy
would have caught up with newDesired. The conditional was removed in
11d7b03e83d403f7b59abea5b4a9dd63e080d2b9.