roll: Add fuzz test for strict removals
While making #442 we discovered that #408 kind of made the fuzz test do
odd things: It wasn't necessarily removing the number of nodes it said
it would be removing.

We'd like to keep the current fuzz test, as it seems to be useful in the
mutable world (where the old RC gets disabled thus doesn't necessarily
remove any nodes).

Now we'd like to add a new test that shows how things will progress in a
world of immutable deployments, where the old RC will indeed remove
exactly the number of nodes specified.