Add a util function for determining RC lock type from consul key
This is useful because there are multiple types of locks that can be
held on an RC: an ownership lock, a mutation lock, and an update
creation lock. The nature of consul is that you can list every key under
a certain root, but if a caller is only intersted in a particular lock
type they have to filter out the lock types they don't care about. This
function makes that easier to do.