Return a specific error type if oldRCSelector is ambiguous
One way to specify the old RC to use when creating an RU is by using a
label selector. If the selector matches more than one RC, an error is
returned.

This commit makes the roll store return a special error type in this
case so that callers can detect this error type and generate more useful
error messages. In Square's application of this function, an error of
this type means that a deploy is already in progress and thus there are
two RCs that match the selector.