add WatchNewWithLockInfo() to rc store
This is currently unused, but can be used to minimize the work done by
the RC farm every time the RC list is updated. The current implmentation
of the RC farm attempts to acquire a lock for each RC, which can be time
consuming when there are 100s of RCs.

WatchNewWithLockInfo() returns a list of RCs embedded in structs that
detail the various locks that are held on the RC. The lock info is
grabbed all at once each time the RC list changes. There may be some
cases where the lock information is stale as a result when the RC update
is being processed, but a subsequent update should see the correct view
of the world and do the right thing