Add p2-rctl-server to integration test
This change expands the role of the integration test to run
a p2-rctl-server. Instead of scheduling a `hello` pod directly
on the host, we instead create a `hello` replication controller
and verify that hello gets correctly scheduled on the host
running the integration test.

One of the new conditions verified is that the `hello` pod gets
labeled with the RC that scheduled it in the label tree. This
exercises the relatively new label aggregation and watch behavior.

This change was relatively painless. The next step is to actually
attempt a rolling update through the same p2-rctl-server.