rollstore: If created an old RC, apply new RC's labels
Since we expect the old RC to be a copy of the new one, then we should
copy the labels as well, except that was not done.

Not copying the labels results in some schemes that assume the label
presence to not be able to find the old RC (since it has no labels).