pcstore implements Watch() and WatchPodCluster()
I'm envisioning a clientside workflow somewhat like the following:

1) An outer loop calls Watch to observe changes to the pod cluster tree,
forking go routines to call WatchPodCluster on new PodClusters
2) These go routines live as long as the PodCluster exists, shutting themselves
down when the PodCluster is deleted.

So, if a pod cluster goes away and comes back it will be observed by the outer
loop and get a watching goroutine fairly expediently. I'm not seeing anything
undesirable about this, but I'm happy to be wrong.