Implement build manifest-based artifact verification.
This change introduces an alteration to the signature of pod.Install(),
requiring an implementation of auth.ArtifactVerifier be passed. This change
also introduced the requisite configuration for p2-preparer to select
an artifact verification strategy.

Some binaries that do installations directly were hardcoded with an
auth.NoopVerifier() for the time being. These will be upgraded in subsequent work.