Only call `SyncCluster` if pods or cluster change.
`SyncCluster` would previously be called on every return from the
label watch. Clients were implementing their own variation of rate
limiting and throttling to deal with the flow of calls. This change
reduces the frequency of unnecessary calls to `SyncCluster`.

`SyncCluster` is still expected to be idempotent. It is still guaranteed
to be called (possibly redundantly) when the watch against the pcstore
expires.