Collect metrics on SyncCluster() time
For certain syncer types, it may take non-negligible amounts of time to
sync a pod cluster. This commit adds support for providing a metrics
registry to the pod cluster store. If a registry is set, a metric will
be added for pod cluster sync time and data points will be collected.