Fix a bug where a label watch wouldn't be started for some pod clusters
pcstore.WatchAndSync() calls the concrete syncer's GetInitialClusters()
function in order to notify the syncers of pod cluster deletions that
occurred when the syncer wasn't running.

However, handlePCUpdates only begins a label query on the pod cluster's
pod selector if something has changed, and there was no apparent change
since we were seeding the data with the GetInitialClusters() call.

This commit fixes that by only seeding the pod cluster changes with the
ID of the pod cluster if it was returned in GetInitialClusters().

The behavior change of this is that a watch will be started for every
pod cluster, and a SyncCluster() call will also be called once per pod
cluster even if the view of the world hasn't changed, perhaps resulting
in some performance degradation when the syncer starts up.