Extract artifact downloading code from Launchable struct
This change is in anticipation of supporting new artifact downloading
schemes. Currently, every launchable stanza must have the Location field
set, and the artifact is fetched from that URL, and there are magic
suffixes that can be added to the URL to get the artifact's manifest and
the manifest's signature for artifact verification.

The new anticipated scheme will result in a URL that instead of
returning an artifact, returns a JSON structure that gives the URL to
each of the artifact, manifest, and signature so that magic suffixes do
not need to be used. This comes along with an increase in complexity to
the artifact download process that cannot be hidden behind the
uri.Fetcher interface as it was prior to this commit.

There should be no behavior changes associated with this commit, things
are just restructured with future implementations of artifact.Downloader
in mind.

Some byproduct changes:

1) Location no longer exists on the launchable struct anymore, as the
downloading code is decoupled from it via the artifact.Downloader
interface.
2) As a result of 1), the install directory is no longer determined via
path.Basename(hl.Location), which was a brittle process anyway as it
asusmes that the end of the URL is suitable for a directory name.
Instead, the directory name is constructed using
<launchable_id>_<version>
3) As a result of 2), a new Version key was added to the Launchable
struct. This version is still extracted from the Location field on the
LaunchableStanza, which is considered brittle as it makes assumptions
about the structure of the URL. When the new artifact downloading scheme
is implemented, this will be easy to accomplish as we anticipate a
first-class "version" key in the launchable stanza.
4) Since the launchable URL is no longer used to determine the name of
the install directory, some test code and file paths to some fixtures
needed to change.
5) The function signature of (Pod) Install(manifest pods.Manifest) error changed
to (Pod) Install(manifest pods.Manifest, verifier auth.ArtifactVerifier)
error as it now needs to construct an artifact.Downloader which requires
an artifact verifier. The code in this function is a little awkward
since it needs to both assemble Launchable{} structs (since the
launchable struct knows the desired destination directory for the
installation) as well as refer to the Location on the launchable stanza.
These two concepts are linked through the launchable ID. There may be a
way to simplify this by extracting the Install() function out of
Launchable completely, but that will be done in a future commit.