Optimizations to outstanding connection counts in pkg/replication.
`pkg/replication` is primarily used by human operators attempting to
deploy many pods simultaneously (in the absence of daemon sets, which
are being actively worked on). For this reason, we want to be considerate
of the number of open file descriptors used by the replication package
at any given time.

This PR switches reality observations into periodic polls (20 + rand(10) seconds)
and changes health to use a single service-wide aggregation. Previously on my
Macbook I was unable to execute any deploy with a parallelism factor over 100. With
this change I was able to perform a deployment at a 200 node parallelism factor without
issue.