Set up artifact verification types in anticipation of discovery
Motivation:
The discovery artifact verification scheme will work as
follows:

a GET /discover (with relevant query params) will return a JSON
structure containing a URL from which the artifact may be downloaded, as
well as relevant verification files. An example might be:

{
  "url": https://registry.com/artifact.tar.gz",
  "manifest": https://registry.com/artifact.tar.gz.manifest",
  "manifest_signature": https://registry.com/artifact.tar.gz.manifest.gz"
}

In which case the (to be implemented) DiscoveryDownloader would need to
use the VerifyManifest artifact verication scheme to validate the
artifact.

Prior to this commit, the auth.ArtifactVerifier interface took two
arguments: a local copy of the downloaded artifact and the url it came
from, which was used to deduce the paths to the other needed files using
magic suffixes.

In the world of using the discover mechanism, magic suffixes may not be
used, so the different artifact schemes cannot be hidden behind such a
simple interface.

As a result, this commit provides separate interfaces for
BuildManifestVerifier and BuildVerifier. A future implementation of
artifact.Downloader that needs to decide artifact verification scheme at
download time will make use of these.

Furthermore, this commit renames auth.ArtifactVerifier to
auth.LocationVerifier, with the same interface as before to minimize
code churn. No existing behavior is expected to have changed.

Lastly, the CompositeVerifier type (which implements
auth.LocationVerifier) does the work of inferring the path to
verification files using the magical paths to maintain the old
functionality. The BuildManifestVerifier and BuildVerifier types require
those locations to be passed in explicitly.