Modify auth.ArtifactVerifier interface to take VerificationData struct
We're planning to support two methods of specifying the location from
which a launchable may be downloaded:

1) using the "location" field in LaunchableStanza
2) using the "version" field in LaunchableStanza. This method is
currently not implemented to keep this commit smaller.

Previously all artifact downloads used the "location" field, and when
artifact verification is desired, the supplementary files needed to
verify the artifact (such as manifest and signatures) were inferred from
the artifact location by adding some special suffixes.

When the "version" scheme is used, an artifact registry will instead be
queried using the artifact metadata under the "version" key (as well as
other data) and the location of the artifact will be returned along with
the locations of the supplementary verification files. As a result, this
commit pulls the suffix-using logic out of artifact verification code.

Now the artifact package has a new function that takes a
LaunchableStanza and returns the location of the launchable and a
VerificationData struct that can be used for artifact verification.
Querying the artifact registry when using a "version" scheme is not yet
implemented.