Move the Manifest type from pkg/pods to pkg/manifest
This change allows the LaunchableStanza type to move out of the
pkg/types package which is a dumping ground for commonly imported types
that result in import cycles when in a more "logical" location.

The immediate motivation was the Pod type (in pkg/pods) having a
reference to the pkg/auth (ArtifactVerifier) type, and the artifact
verifier type having a reference to pkg/pods.LaunchableStanza, resulting
in it moving to the types package.