Adds convenience functions for marshaling RC and PC types as JSON
Due to the fact that the RC and PC structs contain custom types that do
not marshal to JSON automatically, representing them as JSON involves
going through an intermediary type named RawRC and RawPC, respectively.
This commit adds convenience functions for converting an RC or PC to the
corresponding json-friendly type.