Move artifact installation code out of launchables into Pod
Much of the process of installing different launchable types is
identical: it involves fetching a .tar.gz file from somewhere and
unpacking it to a specific directory. Different lauchable types may have
some post-processing setup which can now be done in a PostInstall()
function.

Furthermore, as we support multiple ways of downloading an artifact, it
makes sense to centralize this logic in the pods package rather than
having to duplicate it for each launchable type.

In addition, many parts of the pod setup process require access to the
Launchable struct, and needing to initialize download-specific structs
like an artifact verifier and a downloader is superfluous for these
cases.