Support fetching artifact and verification data locations from registry
This commit adds the ability to download an artifact using the new
"version" scheme in the launchable stanza. Instead of specifying the
exact location for an artifact to be downloaded from and then guessing
the location of the various artifact integrity verification files using
magic suffixes, the launchable's ID along with explicit Version and Tags
can be used to ask an artifact registry for the location of the artifact
and its associated files.