Add better checks when daemon set notices a change
Daemon sets now now takes a pointer to ds fields and double checks if it
is a valid signal

This was made so that the channels take a daemon set pointer so that
you can double check the uuid of the daemon set in case a mistake was
made.

Also, it is possible to get nil values on the channel, so an ok return
value is also taken in during a receive from the channel so that we know
to also quit the go routine. This is done because cases in the select
statements are non-deterministic if more than one case is satisfied

The logging has also been reflected to take in this change and produce
better error messages.

Also changed replaced get with was in some log messages