Make version detection from location URL stricter
Prior to this commit, the version detection code would do so incorrectly
on some artifact names. Now the regex enforces that the version is
a 40 character hexadecimal string with an optional hyphenated suffix.

This resulted in some bad install directory names. For example, a
launchable with id "foo_launchable.tar.gz" would be installed in a
directory named "foo_launchable_launchable/" because the launchable ID
was concatenated with the supposed version "launchable". The desired
behavior would have been for versionFromLocation() to return an error
and the ID of "foo_launchable" alone would be used to determine the
directory name.