Add support for the "entry_points" field in launchable stanza
This field is only used by hoist launchables, and enumerates the path to
a file or directory of files that should be used as entry points for
runit services. If there are no entry points enumerated, a default value
of bin/launch is used which matches the behavior of all hoist
launchables prior to this commit.

A constraint is enforced that no two runit services may have the same
name, which translates to entry points necessarily having differening
file names. For example, a manifest specifying ["bin/launch",
"bin/start"] for entry point is invalid if bin/start and
bin/launch/start are files in the tarball because those would result in
runit services with the same name.